pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "femo"
dynamic = ["version"]
description = "Package for analysis of Fetal Movement data"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["machine-learning", "deep-learning", "sensor-fusion", "multi-modal", "fetal-movement"]
authors = [
  {name = "M A Batin", email = "leonhsn18@gmail.com"},
]
maintainers = [
  {name = "M A Batin", email = "leonhsn18@gmail.com"}
]
classifiers = [
  "Programming Language :: Python :: 3.10",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Microsoft :: Windows",
]

# Required dependencies ------------------------------------------------------------------------------------------------
dependencies = [
  "python-dotenv==1.0.1",
  "boto3>=1.26.118",
  "botocore>=1.29.118",
  "pyyaml>=6.0.2",
  "tqdm==4.66.5",
  "pandas==2.2.0",
  "PyWavelets==1.7.0",
  "scikit-learn==1.3.0",
  "scikit-image==0.24.0",
  "tensorflow==2.15.1",
  "xgboost==2.1.1",
  "openpyxl==3.1.5",
  "matplotlib==3.9.2",
  "tsfel==0.1.9"
]

# Optional dependencies ------------------------------------------------------------------------------------------------
[project.optional-dependencies]
dev = [
    "ipython",
    "pytest",
    "pytest-cov",
    "ruff",
    "typing_extensions==4.12.2"
]
sagemaker = [
    "sagemaker==2.232.0",
    "fastapi==0.103.2",
    "uvicorn==0.32.0",
    "gevent==24.10.3",
    "gunicorn==23.0.0"
]

[project.urls]
"Source" = "https://github.com/MAIMLab-DU/FeMo_Analysis.git"
"Changelog" = "https://github.com/MAIMLab-DU/FeMo_Analysis/releases"

# Tools settings -------------------------------------------------------------------------------------------------------
[tool.setuptools]  # configuration specific to the `setuptools` build backend.
packages = { find = { where = ["."], include = ["femo", "femo.*"] } }
package-data = { "femo" = ["../tests/*.py"] }

[tool.setuptools.dynamic]
version = { attr = "femo.__version__" }

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::FutureWarning"
]
norecursedirs = [".git", "dist", "build"]

[tool.ruff]
line-length = 120